+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.3 - releases
 
 ### Features
+ - Add parameter in route
  - Add class based callback support
  - Add matchRoutes method
  - Add `get`, `post`, `put`, `delete` method for their corresponding HTTP method